Add : Assignment 2 - mock data를 활용하여 여러 개의 피드  구현
