+++ src/pages/jaewonyang/Main/Comment/CommentJaeWon.scss
@@ -1,4 +1,4 @@
-.Comment {
+.CommentJaeWon {
   padding-left: 0;
 
   li {
