+++ src/pages/jaewonyang/Main/Feed/FeedJaeWon.scss
@@ -1,0 +1,112 @@
+.FeedJaeWon {
+  width: 400px;
+  height: auto;
+  border: 1px solid lightgray;
+  background-color: white;
+  margin-bottom: 20px;
+
+  header {
+    height: 50px;
+    display: flex;
+    justify-content: left;
+    align-items: center;
+    position: relative;
+
+    img {
+      width: 30px;
+      height: 30px;
+      border-radius: 50%;
+      border: 2px solid hotpink;
+      margin: 0 10px;
+      padding: 2px;
+    }
+
+    span {
+      font-size: 14px;
+      font-weight: 700;
+    }
+
+    i {
+      position: absolute;
+      right: 20px;
+    }
+  }
+
+  article {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+
+    img {
+      width: 100%;
+      height: 400px;
+    }
+
+    .article__icons {
+      background-color: white;
+      height: 40px;
+      display: flex;
+      position: relative;
+      width: 100%;
+
+      button {
+        font-size: 24px;
+        outline: none;
+        border: none;
+        background-color: transparent;
+        align-items: center;
+        margin: 0 2px;
+        color: gray;
+
+        .fa-heart {
+          color: #ed4956;
+        }
+
+        .fa-bookmark {
+          position: absolute;
+          top: 10px;
+          right: 15px;
+        }
+      }
+    }
+
+    .article__like {
+      padding-left: 10px;
+      display: flex;
+      align-items: center;
+      justify-content: left;
+      padding: 0;
+      font-size: 16px;
+      margin-bottom: 8px;
+
+      img {
+        width: 20px;
+        height: 20px;
+        border: 1px solid lightgray;
+        border-radius: 50%;
+        margin: 0 4px;
+      }
+
+      span:first-child {
+        font-weight: 600;
+      }
+    }
+
+    .article__content {
+      display: flex;
+
+      span:first-child {
+        font-weight: 600;
+        font-size: 16px;
+        margin-left: 10px;
+        margin-right: 5px;
+      }
+
+      span:nth-child(2) {
+        span {
+          color: gray;
+        }
+      }
+    }
+  }
+}
