+++ src/pages/jaewonyang/Main/MainJaeWon.scss
@@ -2,6 +2,8 @@
   margin: auto;
   background-color: #fafafa;
   height: 800px;
+  display: flex;
+  flex-direction: column;
 
   .main {
     display: flex;
@@ -9,158 +11,9 @@
     margin-top: 50px;
     justify-content: space-evenly;
 
-    .main__left {
-      width: 400px;
-      height: 650px;
-      border: 1px solid lightgray;
-      background-color: white;
-
-      header {
-        height: 50px;
-        display: flex;
-        justify-content: left;
-        align-items: center;
-        position: relative;
-
-        img {
-          width: 30px;
-          height: 30px;
-          border-radius: 50%;
-          border: 2px solid hotpink;
-          margin: 0 10px;
-          padding: 2px;
-        }
-
-        span {
-          font-size: 14px;
-          font-weight: 700;
-        }
-
-        i {
-          position: absolute;
-          right: 20px;
-        }
-      }
-
-      article {
-        display: flex;
-        flex-direction: column;
-        align-items: flex-start;
-
-        img {
-          width: 100%;
-          height: 400px;
-        }
-
-        .article__icons {
-          background-color: white;
-          height: 40px;
-          display: flex;
-          position: relative;
-          width: 100%;
-
-          button {
-            font-size: 24px;
-            outline: none;
-            border: none;
-            background-color: transparent;
-            align-items: center;
-            margin: 0 2px;
-            color: gray;
-
-            .fa-heart {
-              color: #ed4956;
-            }
-
-            .fa-bookmark {
-              position: absolute;
-              top: 10px;
-              right: 15px;
-            }
-          }
-        }
-
-        .article__like {
-          padding-left: 10px;
-          display: flex;
-          align-items: center;
-          justify-content: left;
-          padding: 0;
-          font-size: 16px;
-          margin-bottom: 8px;
-
-          img {
-            width: 20px;
-            height: 20px;
-            border: 1px solid lightgray;
-            border-radius: 50%;
-            margin: 0 4px;
-          }
-
-          span:first-child {
-            font-weight: 600;
-          }
-        }
-
-        .article__content {
-          display: flex;
-
-          span:first-child {
-            font-weight: 600;
-            font-size: 16px;
-            margin-left: 10px;
-            margin-right: 5px;
-          }
-
-          span:nth-child(2) {
-            span {
-              color: gray;
-            }
-          }
-        }
-      }
-
-      footer {
-        width: 400px;
-        background-color: white;
-
-        .footer__time {
-          color: gray;
-          margin-left: 10px;
-        }
-
-        form {
-          display: flex;
-          justify-content: space-between;
-          width: 100%;
-          height: 100%;
-          position: relative;
-
-          input {
-            outline: none;
-            border: none;
-            width: 100%;
-            height: 25px;
-            position: relative;
-            left: 0;
-            padding: 4px 0 4px 10px;
-          }
-
-          .comment__button {
-            position: absolute;
-            top: 8px;
-            right: 8px;
-            background-color: transparent;
-            border: none;
-            outline: none;
-            color: #bfe0fd;
-          }
-
-          .active {
-            color: #0095f6;
-          }
-        }
-      }
+    .feed {
+      display: flex;
+      flex-direction: column;
     }
   }
 }
